+++ src/Console/Application/Theme/ThemeInterface.php
@@ -9,12 +9,9 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 interface ThemeInterface
 {
-    public function resetOutput(OutputInterface $output, ProcessArray $processes): self;
+    public function outputStart(OutputInterface $output, ProcessArray $processes): self;
 
-    public function outputProcessesState(
-        OutputInterface $output,
-        ProcessArray $processes
-    ): self;
+    public function outputProcessesState(OutputInterface $output, ProcessArray $processes): self;
 
     public function outputSummary(OutputInterface $output, ProcessArray $processes): self;
 }