+++ changelog.md
@@ -1,5 +1,7 @@
 ### master
 
+### [0.8.2](../../compare/0.8.1...0.8.2) - 2022-03-15
+
 - Fix `SummaryTheme` processes count color
 
 ### [0.8.1](../../compare/0.8.0...0.8.1) - 2022-03-15